+++ thrift/src/main/scala/com/twitter/finatra/thrift/ThriftServer.scala
@@ -1,12 +1,15 @@
 package com.twitter.finatra.thrift
 
 import com.twitter.conversions.time._
-import com.twitter.finagle.{ListeningServer, ThriftMux}
+import com.twitter.finagle.{NullServer, ListeningServer, ThriftMux}
 import com.twitter.finatra.logging.modules.Slf4jBridgeModule
 import com.twitter.finatra.thrift.routing.ThriftRouter
 import com.twitter.inject.annotations.Lifecycle
 import com.twitter.inject.server.{PortUtils, TwitterServer}
 import com.twitter.util.Await
+
+/** AbstractThriftServer for usage from Java */
+abstract class AbstractThriftServer extends ThriftServer
 
 trait ThriftServer extends TwitterServer {
 
@@ -25,7 +28,8 @@ trait ThriftServer extends TwitterServer {
   private val thriftAnnounceFlag = flag[String]("thrift.announce", "Address for announcing Thrift server")
 
   /* Private Mutable State */
-  private var thriftServer: ListeningServer = _
+
+  private var thriftServer: ListeningServer = NullServer
 
   /* Abstract */
 
@@ -52,8 +56,11 @@ trait ThriftServer extends TwitterServer {
         ThriftMux.server
           .withLabel(thriftServerNameFlag()))
 
-    thriftServer =
-      thriftServerBuilder.serveIface(thriftPortFlag(), router.filteredService)
+    thriftServer = router.services.service.map { service =>
+      thriftServerBuilder.serve(thriftPortFlag(), service)
+    }.getOrElse {
+      thriftServerBuilder.serveIface(thriftPortFlag(), router.services.serviceIface)
+    }
     onExit {
       Await.result(
         thriftServer.close(thriftShutdownTimeoutFlag().fromNow))
@@ -69,6 +76,19 @@ trait ThriftServer extends TwitterServer {
 
   /* Protected */
 
+  /**
+   * This method allows for further configuration of the thrift server for parameters not exposed by
+   * this trait or for overriding defaults provided herein, e.g.,
+   *
+   * override def configureThriftServer(server: ThriftMux.Server): ThriftMux.Server = {
+   *   server
+   *     .withResponseClassifier(...)
+   *     .withMaxReusableBufferSize(...)
+   * }
+   *
+   * @param server - the [[com.twitter.finagle.ThriftMux.Server]] to configure.
+   * @return a configured ThriftMux.Server.
+   */
   protected def configureThriftServer(server: ThriftMux.Server): ThriftMux.Server = {
     server
   }